+++ apps/frontend/src/pages/overview/libs/components/add-course-modal/add-course-modal.tsx
@@ -1,4 +1,5 @@
 import { Courses, Input, Loader, Modal } from "~/libs/components/components.js";
+import { EMPTY_ARRAY_LENGTH } from "~/libs/constants/constants.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import { initDebounce } from "~/libs/helpers/helpers.js";
 import {
@@ -32,13 +33,16 @@ type Properties = {
 
 const AddCourseModal: React.FC<Properties> = ({ onClose }: Properties) => {
 	const dispatch = useAppDispatch();
-	const { courses, isLoading, vendors } = useAppSelector((state) => {
-		return {
-			courses: state.courses.searchedCourses,
-			isLoading: state.courses.searchDataStatus === DataStatus.PENDING,
-			vendors: state.vendors.vendors,
-		};
-	});
+	const { courses, isLoading, recommendedCourses, vendors } = useAppSelector(
+		(state) => {
+			return {
+				courses: state.courses.searchedCourses,
+				isLoading: state.courses.searchDataStatus === DataStatus.PENDING,
+				recommendedCourses: state.courses.recommendedCourses,
+				vendors: state.vendors.vendors,
+			};
+		},
+	);
 	const { control, errors, handleSubmit, setValue } = useAppForm({
 		defaultValues: DEFAULT_SEARCH_COURSE_PAYLOAD,
 		mode: "onChange",
@@ -56,6 +60,12 @@ const AddCourseModal: React.FC<Properties> = ({ onClose }: Properties) => {
 	): void => {
 		void dispatch(
 			courseActions.getAll({
+				search: filterFormData.search,
+				vendorsKey: getVendorsFromForm(filterFormData.vendors),
+			}),
+		);
+		void dispatch(
+			courseActions.getRecommended({
 				search: filterFormData.search,
 				vendorsKey: getVendorsFromForm(filterFormData.vendors),
 			}),
@@ -84,6 +94,8 @@ const AddCourseModal: React.FC<Properties> = ({ onClose }: Properties) => {
 			handleDebouncedSearchCourses.clear();
 		};
 	}, [handleDebouncedSearchCourses]);
+
+	const hasCourses = courses.length > EMPTY_ARRAY_LENGTH;
 
 	return (
 		<Modal isOpen onClose={onClose}>
@@ -124,7 +136,21 @@ const AddCourseModal: React.FC<Properties> = ({ onClose }: Properties) => {
 						{isLoading ? (
 							<Loader color="orange" size="large" />
 						) : (
-							<Courses courses={courses} onAddCourse={handleAddCourse} />
+							<>
+								<Courses courses={courses} onAddCourse={handleAddCourse} />
+								{hasCourses && (
+									<div className={styles["recommended-courses"]}>
+										<h2 className={styles["courses-title"]}>
+											Recommended Courses
+										</h2>
+
+										<Courses
+											courses={recommendedCourses}
+											onAddCourse={handleAddCourse}
+										/>
+									</div>
+								)}
+							</>
 						)}
 					</div>
 				</div>